% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorCluster.R
\name{colorCluster}
\alias{colorCluster}
\title{Colors an hclust object according to a vector of conditions}
\usage{
colorCluster(hclus, condition, ce)
}
\arguments{
\item{hclus:}{an hclust object obtained with the hclust function}

\item{condition:}{a numeric vector of the length=numb of samples, each number is a condition(ex: condition=c(1,1,1,1,2,2,2,2))}

\item{Ce:}{cex parameter that is set in function of the length of the characters of the vector}
}
\value{
returns the cluster with colored leafs
}
\description{
Colors an hclust object according to a vector of conditions
}
