% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeHeatmap.R
\name{makeHeatmap}
\alias{makeHeatmap}
\title{Function to make heat map from a numeric matrix using pheatmap package}
\usage{
makeHeatmap(
  matrix,
  resultsDir = NULL,
  fileName = NULL,
  fmtPlot = "",
  title = "",
  scaleColors = c("blue", "white", "red"),
  clusterRow = TRUE,
  clusterCol = TRUE,
  clustMethod = "ward.D2",
  clustDistCols = "correlation",
  scale = "row",
  annot = NA,
  annotColors = NA,
  showRowNames = TRUE,
  fontSizeRows = 10,
  showRowCluster = FALSE,
  ...
)
}
\arguments{
\item{matrix}{Numeric matrix with the intensity values to plot, with row.names (usually genes) and col.names (samples)}

\item{resultsDir}{Output directory of the plot. Default = NULL}

\item{fileName}{Name of the output file, without extension. Default = NULL}

\item{fmtPlot}{Format for the image file, "png" or "pdf". If none specified image will pop up in R session. Default = ""}

\item{title}{Title of the plot. Default = ""}

\item{scaleColors}{Vector with three colors to compose the color scale. Default = c("blue","white","red")}

\item{clusterRow}{Whether to cluster rows. Equivalent to parameter cluster_row of pheatmap. Default = TRUE}

\item{clusterCol}{Whether to cluster columns. Equivalent to parameter cluster_col of pheatmap. Default = TRUE}

\item{clustMethod}{Clustering method. Default "ward.D2". Values given by parameter clustering_method of pheatmap function. Default = "ward.D2"}

\item{clustDistCols}{Distance for clustering columns. Values given by parameter clustering_distance_cols of pheatmap function. Default = "correlation"}

\item{scale}{Character indicating if the values should be centered and scaled. Values given by parameter scale of pheatmap function. Default = "row"}

\item{annot}{Dataframe that specifies the annotations shown for columns. rownames of the data.frame must be the colnames of the matrix object. Default = NA}

\item{annotColors}{List with as many elements as conditions. Each element must be a vector with the color assignment for each category of the condition. Default = NA (assigned internally)}

\item{fontSizeRows}{Size of row labels. Computed depending on the number of rows of matrix. Default = 10}

\item{showRowCluster}{Whether to show the tree of the row clustering. Default = FALSE}

\item{...}{Additional parameters of pheatmap}

\item{showRownames}{Indicates whether you want to show rownames or not. Each row defines the features for a specific row. A maximum of 150 genes is possible, if this number is exceeded the parameter will automatically be turned off. Default = TRUE}
}
\value{
The heatmap plot is created in the "resultsDir" with the name "fileName"
}
\description{
Function to make heat map from a numeric matrix using pheatmap package
}
