% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makePCA.R
\name{makePCA3D}
\alias{makePCA3D}
\title{Creates a PCA 3D from a numerical matrix}
\usage{
makePCA3D(
  estimates,
  resultsDir = NULL,
  fileName = NULL,
  fmtPlot = "pdf",
  title = NULL,
  conditions = NULL,
  colors = NULL,
  dist = 2,
  dim = 3,
  legend.pos = "bottomright"
)
}
\arguments{
\item{estimates}{Numeric matrix with the intensity values to plot.}

\item{resultsDir}{Output directory. Default = NULL}

\item{fileName}{Name of the output file, without extension. Default = NULL}

\item{fmtPlot}{Format for the image file, "png" or "pdf" (default). If none specified image will pop up in R session}

\item{title}{Title for each plot. Default = NULL}

\item{conditions}{Vector with the different conditions. Default = NULL}

\item{colors}{Vector with the colors assigned to each condition (in order of the unique(conditions)). Default = NULL}

\item{dist}{Distance from labels to dots. Default = 2}

\item{dim}{Dimensions of the PCA plot. Possible values are 2 or 3. Default = 3}

\item{legend.pos}{position of the legend. Default "bottomright"}
}
\value{
PCA image
}
\description{
Creates a PCA 3D from a numerical matrix
}
