% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeVenn.R
\name{makeVenn}
\alias{makeVenn}
\title{Creates a Venn diagram based on several gene lists}
\usage{
makeVenn(
  listGenes,
  resultsDir = NULL,
  fileName = NULL,
  fmtPlot = "",
  title = NULL,
  mkExcel = TRUE,
  colors = NULL,
  labelSize = 6,
  setSize = 5,
  percentage = FALSE,
  wid = 8,
  hei = 8,
  res = 400,
  ...
)
}
\arguments{
\item{listGenes}{Named list with gene vectors, with 2 to 4 elements. Names should be short strings of up to 10 characters}

\item{resultsDir}{Output directory. Default = NULL}

\item{fileName}{Name of the output file, without extension. Default = NULL}

\item{fmtPlot}{Format for the image file, "png" or "pdf". If none specified image will pop up in R session. Default = ""}

\item{title}{Title for each plot. Default = NULL}

\item{mkExcel}{Whether to generate an excel with gene intersection and specific sections. Default = TRUE}

\item{colors}{Vector of colors to paint the different main conditions of the Venn diagram. Intersecting sections will be colored with color overlap. If none specified, pre-defined default colors will be used.}

\item{labelSize}{Text size of the name label of each set. Equivalent to set_name_size from ggvenn. Default = 6}

\item{setSize}{Text size for the set number contents. Equivalent to text_size from ggvenn. Default = 5}

\item{percentage}{Whether to show the percentage of genes for each set. Default = FALSE}

\item{wid}{Plot parameter: width of the device (in inches). Default = 8}

\item{hei}{Plot parameter: height of the device (in inches). Default = 8}

\item{res}{Plot parameter: nominal resolution in ppi. Default = 400}

\item{...}{Additional parameters to add to ggvenn function}
}
\value{
Venn diagram plot and list of genes in each set (in excel format) if specified
}
\description{
Creates a Venn diagram based on several gene lists
}
