% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeVolcano.R
\name{makeVolcano}
\alias{makeVolcano}
\title{Function to make volcano plot from a topTable extracted from limma}
\usage{
makeVolcano(
  topTable,
  resultsDir = volcanoDir,
  fileName = NULL,
  fmtPlot = "",
  title = NULL,
  scaleColors = c("blue", "grey", "red"),
  yVal = "p.adjust",
  p.val = NULL,
  p.adj = 0.05,
  thres.logFC = 1,
  coefs = fit.main$coefficients,
  topGenes = 20,
  maxOverlaps = 25,
  geneLabel = "Geneid",
  sortLabel = "p-value",
  annotFile = NULL,
  interactive = FALSE,
  ...
)
}
\arguments{
\item{resultsDir}{Output directory. Default = volcanoDir}

\item{fileName}{Name of the output file, without extension. Default = NULL}

\item{fmtPlot}{Format for the image file, "png" or "pdf". If none specified ("") image will pop up in R session. Default = ""}

\item{title}{Title for each plot. Default = NULL}

\item{scaleColors}{Vector with three colors to compose the color scale. Default = c("blue","grey","red")}

\item{yVal}{Value to plot on the y axis. Possible values are p.value or p.adjust. Values will be transformed to minus(log2(yVal)). Default = p.adjust}

\item{p.val}{P-value cutoff for topTable. Default = NULL}

\item{p.adj}{Adjusted p-value cutoff for topTable. Default = 0.05}

\item{thres.logFC}{LogFC cutoff for topTable. Default = 1}

\item{coefs}{Coefficients of fit.main results. Will be used for the limits of x in the volcano plot. Default = fit.main$coefficients}

\item{topGenes}{Number of genes to label. Default = 20}

\item{geneLabel}{Variable with the label of genes to be shown in plot. Default = Geneid}

\item{sortLabel}{Sorting criteria for the label highlight, either "p-value" or "logFC". Will show in plot the labels of the top most significant or most different in absolute fold change, respectively. Default = p-value}

\item{annotFile}{Dataframe that includes the equivalence of Geneid and geneLabel, if geneLabel is specified. Default = NULL}

\item{interactive}{Prints an interactive plot showing the gene ids when hovering the points. Currently only working when geneLabel is "Geneid" and in session (fmtPlot = "")}

\item{dataV}{data.frame as the output obtained through function topTable() in limma package}
}
\value{
The plot is created in the "resultsDir" with the name "fileName" or in session if the format is not specified.
}
\description{
Function to make volcano plot from a topTable extracted from limma
}
