% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeCluster.R
\name{manyClusters}
\alias{manyClusters}
\title{Creates a pdf with different dendrograms (hierarchical clusters) combining a distance and a linkage method}
\usage{
manyClusters(
  estimates,
  use.cor = "pairwise.complete.obs",
  cor.method = "pearson",
  resultsDir = NULL,
  fileName = "manyClusters",
  fmtPlot = "pdf",
  title = NULL,
  conditions = NULL,
  colors = NULL,
  wid = NULL,
  hei = NULL,
  res = NULL,
  ce = NULL
)
}
\arguments{
\item{estimates}{Numeric matrix with the intensity values to plot. The matrix should have samples as columns and genes/transcripts/exons as rows}

\item{use.cor}{Used for correlation distance: method for computing covariances in the presence of missing values. Default = "pairwise.complete.obs"}

\item{cor.method}{Correlation coefficient (or covariance) to be computed. Default = "pearson"}

\item{resultsDir}{Output directory. Default = NULL}

\item{fileName}{name of the output file, without extension. Default = "manyClusters"}

\item{fmtPlot}{Format for the image file. If none specified images will be generated in R session. Default = "pdf"}

\item{title}{Title for each plot. Default = NULL}

\item{conditions}{Vector with the different conditions. Default = NULL}

\item{colors}{Vector with the colors assigned to each condition. Default = NULL}

\item{wid}{Plot parameter: width of the device (in pixels). Internal default (3500) will be used if not specified. Default = NULL}

\item{hei}{Plot parameter: height of the device (in pixels). Internal default (3500) will be used if not specified. Default = NULL}

\item{res}{Plot parameter: nominal resolution in ppi. Internal default (400) will be used if not specified. Default = NULL}

\item{ce}{Plot parameter: character expansion is the numerical value giving the amount by which plotting text and symbols should be magnified relative to the default. Will be defined internally according to character length and number of samples if not specified. Default = NULL}
}
\value{
a pdf with the hierarchical clusters
}
\description{
Creates a pdf with different dendrograms (hierarchical clusters) combining a distance and a linkage method
}
