% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeCluster.R
\name{oneCluster}
\alias{oneCluster}
\title{Creates a pdf or png with a dendrogram (hierarchical clusters) combining a distance and a linkage method}
\usage{
oneCluster(
  estimates,
  resultsDir = NULL,
  fileName = NULL,
  fmtPlot = "pdf",
  title = NULL,
  distance = "correlation",
  use.cor = "pairwise.complete.obs",
  cor.method = "pearson",
  clust.method = "ward.D2",
  conditions = NULL,
  colors = NULL,
  wid = NULL,
  hei = NULL,
  res = NULL,
  ce = NULL,
  ...
)
}
\arguments{
\item{estimates}{Numeric matrix with the intensity values to plot}

\item{resultsDir}{Output directory. Default = NULL}

\item{fileName}{name of the output file, without extension. Default = NULL}

\item{fmtPlot}{Format for the image file, pdf or png. If none specified images will be generated in R session. Default = "pdf"}

\item{title}{Title for each plot. Default = NULL}

\item{distance}{Distance measure for the hierarchical clustering. Either correlation or euclidean. Default = "correlation"}

\item{use.cor}{Used for correlation distance: method for computing covariances in the presence of missing values. Default = "pairwise.complete.obs"}

\item{cor.method}{Correlation coefficient (or covariance) to be computed. Default = "pearson"}

\item{clust.method}{Linkage method to cluster samples. Values given by parameter method of hclust function. Default = "ward.D2"}

\item{conditions}{Vector with the different conditions. Default = NULL}

\item{colors}{Vector with the colors assigned to each condition (in order of the unique(conditions)). Default = NULL}

\item{wid}{Plot parameter: width of the device (in pixels). Internal default (3500) will be used if not specified. Default = NULL}

\item{hei}{Plot parameter: height of the device (in pixels). Internal default (3500) will be used if not specified. Default = NULL}

\item{res}{Plot parameter: nominal resolution in ppi. Internal default (400) will be used if not specified. Default = NULL}

\item{ce}{Plot parameter: character expansion is the numerical value giving the amount by which plotting text and symbols should be magnified relative to the default. Will be defined internally according to character length and number of samples if not specified. Default = NULL}

\item{...}{Other elements of plots}
}
\value{
a pdf and/or png with the hierarchical cluster
}
\description{
Creates a pdf or png with a dendrogram (hierarchical clusters) combining a distance and a linkage method
}
