% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setParameters.R
\name{setParameters}
\alias{setParameters}
\title{Estimates the parameters for the different graphics}
\usage{
setParameters(labels)
}
\arguments{
\item{labels}{Vector with the sample names}
}
\value{
a list of the parameters needed to perform graphics
}
\description{
Estimates the parameters for the different graphics
}
