% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeHeatmap.R
\name{makeHeatmap}
\alias{makeHeatmap}
\title{Function to make heat map from a numeric matrix using pheatmap package}
\usage{
makeHeatmap(
  matrix,
  resultsDir = NULL,
  fileName = NULL,
  fmtPlot = "pdf",
  title = NULL,
  scaleColors = c("blue", "white", "red"),
  clustMethod = "ward.D2",
  clustDistCols = "correlation",
  scale = "row",
  annot = NA,
  annotColors = NA,
  showRowNames = TRUE,
  fontSizeRows = 10,
  ...
)
}
\arguments{
\item{matrix:}{Numeric matrix with the intensity values to plot, with row.names and col.names}

\item{resultsDir:}{Output directory. Default = NULL}

\item{fileName:}{name of the output file, without extension. Default = NULL}

\item{fmtPlot:}{Format for the image file, "png" or "pdf" (default). If none specified image will pop up in R session}

\item{scaleColors:}{Vector with three colors to compose the color scale. Default = c("blue","white","red")}

\item{title:}{Title for each plot. Default = NULL}

\item{clustMethod:}{Clustering method. Default "ward.D2". Values given by parameter clustering_method of pheatmap function. Default = "ward.D2"}

\item{clustDistCols:}{Distance for clustering columns. Default "correlation". Values given by parameter clustering_distance_cols of pheatmap function}

\item{scale:}{Character indicating if the values should be centered and scaled. Values given by parameter scale of pheatmap function. efault = "row"}

\item{annot:}{data.frame that specifies the annotations shown for columns. rownames of the data.frame must be the colnames of the matrix object. Default = NA}

\item{annotColors:}{list with as many elements as conditions. Each element must be a vector with the color assignment for each category of the condition. Default = NA (assigned internally)}

\item{showRownames:}{Indicates whether you want to show rownames or not. Each row defines the features for a specific row. Default = TRUE}

\item{fontSizeRows:}{Size of row labels. Computed depending on the number of rows of matrix. Default = 10}
}
\value{
The plot is created in the "resultsDir" with the name "fileName"
}
\description{
Function to make heat map from a numeric matrix using pheatmap package
}
