% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makePCA.R
\name{makePCA}
\alias{makePCA}
\title{Creates a PCA 3D from a numerical matrix}
\usage{
makePCA(
  estimates,
  resultsDir = NULL,
  fileName = NULL,
  fmtPlot = "pdf",
  title,
  conditions = NULL,
  colors = NULL,
  dist = 2,
  dim = 3,
  legend.pos = "bottomright"
)
}
\arguments{
\item{legend.pos}{position of the legend. Default "bottomright"}

\item{estimates:}{Numeric matrix with the intensity values to plot}

\item{resultsDir:}{Output directory}

\item{fileName:}{name of the output file, without extension}

\item{fmtPlot:}{Format for the image file, "png" or "pdf" (default). If none specified image will pop up in R session}

\item{title:}{Title for each plot}

\item{conditions:}{Vector with the different conditions}

\item{colors:}{Vector with the colors assigned to each condition (in order of the unique(conditions)). Default = NULL}

\item{dim:}{Dimensions of the PCA plot. Default = 3. Possible values are 2 or 3}

\item{dist:}{Distance from labels to dots. Default = 2}
}
\value{
PCA image
}
\description{
Creates a PCA 3D from a numerical matrix
}
