% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeVenn.R
\name{makeVenn}
\alias{makeVenn}
\title{Creates a Venn diagram based on several gene lists}
\usage{
makeVenn(
  listGenes,
  resultsDir = NULL,
  fileName = NULL,
  fmtPlot = "pdf",
  title = NULL,
  mkExcel = TRUE,
  colors
)
}
\arguments{
\item{listGenes:}{named list with gene vectors, with 2 to 4 elements. Names should be short strings of up to 10 characters}

\item{resultsDir:}{Output directory. Default = NULL}

\item{fileName:}{name of the output file, without extension. Default = NULL}

\item{fmtPlot:}{Format for the image file, "png" or "pdf" (default). If none specified image will pop up in R session}

\item{title:}{Title for each plot. Default = NULL}

\item{mkExcel:}{whether to generate an excel with gene intersection and specific sections. Default = TRUE}

\item{colors:}{vector of colors to paint the different sections of the Venn diagram}
}
\value{
a file
}
\description{
Creates a Venn diagram based on several gene lists
}
