% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oneCluster.R
\name{oneCluster}
\alias{oneCluster}
\title{Creates a pdf with different dendrograms (hierarchical clusters) combining a distance and a linkage method}
\usage{
oneCluster(
  estimates,
  resultsDir = NULL,
  fileName = NULL,
  fmtPlot = "pdf",
  title = NULL,
  method = "ward.D2",
  conditions = NULL,
  colors = NULL
)
}
\arguments{
\item{estimates:}{Numeric matrix with the intensity values to plot}

\item{resultsDir:}{Output directory. Default = NULL}

\item{fileName:}{name of the output file, without extension. Default = NULL}

\item{fmtPlot:}{Format for the image file, pdf" (default). If none specified images will be generated in R session}

\item{title:}{Title for each plot. Default = NULL}

\item{method:}{linkage method to cluster samples. Values given by parameter method of hclust function. Default = "ward.D2"}

\item{conditions:}{Vector with the different conditions. Default = NULL}

\item{colors:}{Vector with the colors assigned to each condition (in order of the unique(conditions)). Default = NULL}
}
\value{
a pdf and/or png with the hierarchical cluster
}
\description{
Creates a pdf with different dendrograms (hierarchical clusters) combining a distance and a linkage method
}
